Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,22 +1,22 @@
 #include <stdio.h>
 int main(){
-    float x,y;
+    int x,y;
     char z;
     scanf("%f%f",&x,&y);
     scanf(" %c",&z);
     switch(z){
         case '+':
-        printf("%.0f",x+y);
+        printf("%d",x+y);
         break;
         case '-':
-        printf("%.0f",x-y);
+        printf("%d",x-y);
         break;
         case '*':
-        printf("%.0f",x*y);
+        printf("%d",x*y);
         break;
         case '/':
         if(y>0){
-        printf("%.0f",x/y);
+        printf("%d",x/y);
         }else{
             printf("error");
         }
